redpanda: move admin server into separate dir
