+++ src/v/redpanda/admin/server.h
